requirements.txt
fastapi==0.125.0
openai==2.28.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
sqlmodel==0.0.37
uvicorn==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pyyaml==6.0.2
email-validator==2.2.0
